Moved app into pages
