+++ app/src/pages/__init__.py
@@ -1,13 +1,13 @@
-from .page1 import Page1
-from .page2 import Page2
+from .instagram_search import InstagramSearch
+from .file_upload import FileUpload
 from ..utils import Page
 
 from typing import Dict, Type
 
 
 PAGE_MAP: Dict[str, Type[Page]] = {
-    "Page 1": Page1,
-    "Page 2": Page2,
+    "Instagram Search": InstagramSearch,
+    "File Upload": FileUpload,
 }
 
 __all__ = ["PAGE_MAP"]
