+++ app/src/utils.py
@@ -1,7 +1,7 @@
-from dataclasses import dataclass
 from abc import ABC, abstractmethod
 
 import streamlit as st
+import pandas as pd
 
 
 class Page(ABC):
@@ -18,3 +18,15 @@ def add_custom_css():
         """,
         unsafe_allow_html=True,
     )
+
+
+def plot_coords(df: pd.DataFrame):
+    """Plots GPS coordinates on Streamlit map
+
+    Args:
+        df (pd.DataFrame): table of latitudes and longitudes
+    Only works where columns named: "lat", "lng"
+    """
+    # TODO: exception handling
+    lat_lng = df[["lat", "lng"]].dropna()
+    st.map(lat_lng, longitude="lng")  # type: ignore
